+++ doc/version_history.rst
@@ -9,21 +9,30 @@ Version History
 v1.1.0
 ------
 
+* Deprecation warning: do not access the ``reader``, ``writer``, or (in the case of `OneClientServer`) ``server`` attributes.
+  Use the provided read and write methods, instead, and let `OneClientServer` manage its asyncio server.
 * Add `OneClientReadLoopServer`: an abstract subclass of `OneClientServer` that includes a read loop.
 * `BaseClientOrServer`: add ``read_str``, ``write_str``, ``read_json`` and ``write_json`` methods.
 * `BaseClientOrServer` and subclasses:
 
+      This also applies to the `Client` and `OneClientServer` subclasses.
     * Add optional ``encoding`` and ``terminator`` constructor arguments.
     * Make constructor arguments keyword-only.
 
-* Add ``BaseOneClientServerTestCase``: a base clase for unit tests of `OneClientServer` and subclasses.
+* Add `BaseOneClientServerTestCase`: a base clase for unit tests of `OneClientServer` and subclasses.
 * Add new constants:
 
     * ``DEFAULT_LOCALHOST``, the same as deprecated ``LOCAL_HOST``
     * ``DEFAULT_TERMINATOR``, the same as deprecated ``TERMINATOR``.
     * ``DEFAULT_ENCODING``.
 
-* `OneClientServer` and subclasses: make the default host ``DEFAULT_LOCALHOST``.
+* `Client`: fix setting ``done_task`` done.
+
+* `OneClientServer` and subclasses:
+
+    * Provide a default for the ``host`` constructor argument: ``DEFAULT_LOCALHOST``.
+    * Fix setting ``done_task`` done.
+
 * Improve class documentation.
 * Use ts_pre_commit_conf.
 * Remove scons support.
@@ -31,8 +40,8 @@ v1.1.0
 v1.0.1
 ------
 
-* pre-commit: update black to 23.1.0, isort to 5.12.0, mypy to 1.0.0, and pre-commit-hooks to v4.4.0.p
-* ``Jenkinsfile``: modernize do not run as root.
+* pre-commit: update black to 23.1.0, isort to 5.12.0, mypy to 1.0.0, and pre-commit-hooks to v4.4.0.
+* ``Jenkinsfile``: do not run as root.
 
 v1.0.0
 ------